# packages/predictive-models/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Predictive Models Service",
    "description": "A microservice for training and serving predictive models",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/tune/meal-activity": {
      "post": {
        "tags": [
          "meal-activity-tuning"
        ],
        "summary": "Tune Meal Activity",
        "description": "Optimize Carb Ratio and Activity Coefficients using fixed Foundation Baseline.",
        "operationId": "tune_meal_activity_api_v1_tune_meal_activity_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MealActivityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MealActivityResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MealActivityRequest": {
        "properties": {
          "windows": {
            "items": {
              "$ref": "#/components/schemas/TimeWindow"
            },
            "type": "array",
            "title": "Windows"
          },
          "baseline_isf": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Baseline Isf"
          },
          "baseline_basal": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Baseline Basal"
          },
          "current_cr": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Current Cr"
          },
          "current_activity_coeffs": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Activity Coeffs"
          },
          "lambda_l2": {
            "type": "number",
            "title": "Lambda L2",
            "default": 0.2
          },
          "lambda_smooth": {
            "type": "number",
            "title": "Lambda Smooth",
            "default": 0.05
          }
        },
        "type": "object",
        "required": [
          "windows",
          "baseline_isf",
          "baseline_basal",
          "current_cr"
        ],
        "title": "MealActivityRequest"
      },
      "MealActivityResult": {
        "properties": {
          "isf": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Isf"
          },
          "cr": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Cr"
          },
          "basal": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Basal"
          },
          "activity_coefficients": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Activity Coefficients"
          },
          "isf_confidence": {
            "items": {
              "prefixItems": [
                {
                  "type": "number"
                },
                {
                  "type": "number"
                }
              ],
              "type": "array",
              "maxItems": 2,
              "minItems": 2
            },
            "type": "array",
            "title": "Isf Confidence"
          },
          "cr_confidence": {
            "items": {
              "prefixItems": [
                {
                  "type": "number"
                },
                {
                  "type": "number"
                }
              ],
              "type": "array",
              "maxItems": 2,
              "minItems": 2
            },
            "type": "array",
            "title": "Cr Confidence"
          },
          "basal_confidence": {
            "items": {
              "prefixItems": [
                {
                  "type": "number"
                },
                {
                  "type": "number"
                }
              ],
              "type": "array",
              "maxItems": 2,
              "minItems": 2
            },
            "type": "array",
            "title": "Basal Confidence"
          },
          "activity_confidence": {
            "additionalProperties": {
              "prefixItems": [
                {
                  "type": "number"
                },
                {
                  "type": "number"
                }
              ],
              "type": "array",
              "maxItems": 2,
              "minItems": 2
            },
            "type": "object",
            "title": "Activity Confidence"
          },
          "r_squared": {
            "type": "number",
            "title": "R Squared"
          },
          "rmse": {
            "type": "number",
            "title": "Rmse"
          },
          "mae": {
            "type": "number",
            "title": "Mae"
          },
          "windows_analyzed": {
            "type": "integer",
            "title": "Windows Analyzed"
          },
          "total_windows": {
            "type": "integer",
            "title": "Total Windows"
          },
          "stable_windows": {
            "type": "integer",
            "title": "Stable Windows"
          },
          "meal_windows": {
            "type": "integer",
            "title": "Meal Windows"
          },
          "activity_windows": {
            "type": "integer",
            "title": "Activity Windows"
          }
        },
        "type": "object",
        "required": [
          "isf",
          "cr",
          "basal",
          "activity_coefficients",
          "isf_confidence",
          "cr_confidence",
          "basal_confidence",
          "activity_confidence",
          "r_squared",
          "rmse",
          "mae",
          "windows_analyzed",
          "total_windows",
          "stable_windows",
          "meal_windows",
          "activity_windows"
        ],
        "title": "MealActivityResult",
        "description": "Results from meal and activity optimization"
      },
...
      "TimeWindow": {
        "properties": {
          "start": {
            "type": "string",
            "title": "Start"
          },
          "end": {
            "type": "string",
            "title": "End"
          },
          "duration_hours": {
            "type": "number",
            "title": "Duration Hours"
          },
          "glucose_start": {
            "type": "number",
            "title": "Glucose Start"
          },
          "glucose_end": {
            "type": "number",
            "title": "Glucose End"
          },
          "glucose_change": {
            "type": "number",
            "title": "Glucose Change"
          },
          "glucose_readings_count": {
            "type": "integer",
            "title": "Glucose Readings Count"
          },
          "bolus_insulin": {
            "type": "number",
            "title": "Bolus Insulin"
          },
          "basal_insulin_delivered": {
            "type": "number",
            "title": "Basal Insulin Delivered"
          },
          "total_insulin": {
            "type": "number",
            "title": "Total Insulin"
          },
          "insulin_activity": {
            "type": "number",
            "title": "Insulin Activity"
          },
          "carbs_consumed": {
            "type": "number",
            "title": "Carbs Consumed"
          },
          "carb_events_count": {
            "type": "integer",
            "title": "Carb Events Count"
          },
          "carb_absorption": {
            "type": "number",
            "title": "Carb Absorption"
          },
          "activity_steps": {
            "type": "number",
            "title": "Activity Steps",
            "default": 0.0
          },
          "activity_calories": {
            "type": "number",
            "title": "Activity Calories",
            "default": 0.0
          },
          "activity_floors": {
            "type": "number",
            "title": "Activity Floors",
            "default": 0.0
          },
          "activity_heart_rate": {
            "type": "number",
            "title": "Activity Heart Rate",
            "default": 0.0
          },
          "activity_hr_elevation": {
            "type": "number",
            "title": "Activity Hr Elevation",
            "default": 0.0
          },
          "activity_impact": {
            "type": "number",
            "title": "Activity Impact",
            "default": 0.0
          },
          "activity_intensity": {
            "type": "string",
            "title": "Activity Intensity",
            "default": "unknown"
          },
          "hour_of_day": {
            "type": "integer",
            "title": "Hour Of Day"
          },
          "is_stable": {
            "type": "boolean",
            "title": "Is Stable"
          },
          "has_meals": {
            "type": "boolean",
            "title": "Has Meals"
          },
          "has_corrections": {
            "type": "boolean",
            "title": "Has Corrections"
          },
          "data_quality": {
            "type": "object",
            "title": "Data Quality"
          },
          "autosens_ratio": {
            "type": "number",
            "title": "Autosens Ratio",
            "default": 1.0
          },
          "basal_drift": {
            "type": "number",
            "title": "Basal Drift",
            "default": 0.0
          },
          "isolation_confidence": {
            "type": "number",
            "title": "Isolation Confidence",
            "default": 1.0
          },
          "unexplained_residual": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unexplained Residual"
          },
          "total_predicted_impact": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Predicted Impact"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "start",
          "end",
          "duration_hours",
          "glucose_start",
          "glucose_end",
          "glucose_change",
          "glucose_readings_count",
          "bolus_insulin",
          "basal_insulin_delivered",
          "total_insulin",
          "insulin_activity",
          "carbs_consumed",
          "carb_events_count",
          "carb_absorption",
          "hour_of_day",
          "is_stable",
          "has_meals",
          "has_corrections",
          "data_quality"
        ],
        "title": "TimeWindow",
        "description": "Single time window for analysis"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}